🐛 always redirect back to app
In case of an issue the OpenId endpoints should always still redirect
back to the App. The App should be responsible for how to show the error
to the user.

The `JoseError` log error has been slightly improved. It could also fail
if the JWKS endpoint is badly configured or not available.